GUACAMOLE-839: Move nonce-generation service to internal common SSO library.
